requirements.txt
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
Flask-Login==0.5.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gevent==20.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector==2.2.9
numpy==1.18.2
pandas==1.0.3
passlib==1.7.2
psycopg2-binary==2.8.5
pycountry==19.8.18
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.0
WTForms==2.2.1
